+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Releases
 
+- **Breaking-Change**: The `typed_route` decorator can now receive `status_code` parameter as a keyword argument.
+- **Add**: The Field class now accepts the `explode` an `style` parameters to define the serialization 
+  behavior according to the OpenAPI specification. Supported styles are `simple`, `form`, `spaceDelimited`, `pipeDelimited`,
+
 ### v0.2.1 (2025-01-29)
 
 - **Breaking-Change**: `FlaskTypedRoutes` fields can't receive the `multi` parameter in the constructor anymore. Now,